refactor: move merge options in a new crate + update example
