+++ Cargo.toml
@@ -41,6 +41,7 @@ members = [
 	"crates/katana/sync/stage",
 	"crates/katana/tasks",
 	"crates/katana/trie",
+	"crates/macros/merge-options",
 	"crates/metrics",
 	"crates/sozo/scarbext",
 	"crates/sozo/signers",
@@ -130,13 +131,15 @@ torii-server = { path = "crates/torii/server" }
 torii-sqlite = { path = "crates/torii/sqlite" }
 torii-typed-data = { path = "crates/torii/typed-data" }
 torii-mcp = { path = "crates/torii/mcp" }
-torii-derive = { path = "crates/torii/derive" }
 
 # sozo
 sozo-ops = { path = "crates/sozo/ops" }
 sozo-scarbext = { path = "crates/sozo/scarbext" }
 sozo-signers = { path = "crates/sozo/signers" }
 sozo-walnut = { path = "crates/sozo/walnut" }
+
+# macros
+merge-options = { path = "crates/macros/merge-options" }
 
 anyhow = "1.0.89"
 arbitrary = { version = "1.3.2", features = [ "derive" ] }
